% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Alignment.R
\name{addMICSVData}
\alias{addMICSVData}
\title{Attach MI CSV Data to Alignment Object}
\usage{
addMICSVData(object, path)
}
\arguments{
\item{object}{An \code{Alignment} object.}

\item{path}{Path to the MI CSV file.}
}
\value{
Updated \code{Alignment} object with \code{mi_data} slot populated.
}
\description{
Loads raw MI data into the Alignment object (as a data frame).
}
