% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Alignment.R
\name{assertAllAlignmentsEqualLength}
\alias{assertAllAlignmentsEqualLength}
\alias{assertAllAlignmentsEqualLength,Alignment-method}
\title{Check That All Sequences Are Equal Length}
\usage{
assertAllAlignmentsEqualLength(object, ...)

\S4method{assertAllAlignmentsEqualLength}{Alignment}(object)
}
\arguments{
\item{object}{An \code{Alignment} object.}

\item{...}{Additional arguments (unused).}
}
\value{
An integer: sequence length if equal; otherwise, \code{-1}.

An integer: the shared alignment length if consistent, or \code{-1} if inconsistent.
}
\description{
Generic function to check if all aligned sequences in an \code{Alignment} object have the same length.
}
\section{Functions}{
\itemize{
\item \code{assertAllAlignmentsEqualLength(Alignment)}: Checks whether all sequences in the alignment have equal length. Returns the length if equal, or \code{-1} if they differ.

}}
