% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Alignment.R
\name{get_alignment_length}
\alias{get_alignment_length}
\title{Get Alignment Length}
\usage{
get_alignment_length(object)
}
\arguments{
\item{object}{An \code{Alignment} object.}
}
\value{
An integer representing the alignment length, or 0 if empty.
}
\description{
Returns the length of the aligned sequences (including gaps).
}
