% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Alignment.R
\name{get_sequence_names}
\alias{get_sequence_names}
\title{Get Sequence Names}
\usage{
get_sequence_names(object)
}
\arguments{
\item{object}{An \code{Alignment} object.}
}
\value{
A character vector of sequence names.
}
\description{
Returns the sequence names (headers) from the alignment.
}
