% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTools.R
\name{getHistWithLabels}
\alias{getHistWithLabels}
\title{Annotate MI scores by gene regions}
\usage{
getHistWithLabels(iscores, igenes, imidway)
}
\arguments{
\item{iscores}{Data frame of MI scores.}

\item{igenes}{A vector of gene labels (e.g., c("gene1", "gene2", "interface")).}

\item{imidway}{Integer index for gene boundary.}
}
\value{
A grouped tibble by gene.
}
\description{
Annotate MI scores by gene regions
}
