% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getMICovarianceScore3}
\alias{getMICovarianceScore3}
\title{Get MI Covariance Score (Version 3)}
\usage{
getMICovarianceScore3(object, ...)
}
\arguments{
\item{object}{An \code{Alignment} object.}

\item{...}{Additional arguments.}
}
\value{
A numeric mutual information score.
}
\description{
Alternate version of \code{getMICovarianceScore}, get the Covariance Scores based on getScore3 which may include perfectly conserved residues.
}
