% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTools.R
\name{getResidueFrequencies}
\alias{getResidueFrequencies}
\title{Get Residue Frequencies from a Sequence}
\usage{
getResidueFrequencies(sequence)
}
\arguments{
\item{sequence}{A character string representing a protein sequence (e.g., \code{"MKLIVT..."}) using one-letter residue codes.}
}
\value{
A numeric matrix (20x1) where each row corresponds to the count of a standard residue in the input sequence, in the same order as returned by \code{getResidues()}.
}
\description{
Computes the frequency of each standard amino acid residue in a given input sequence.
}
\examples{
getResidueFrequencies("ACDEFGHIKLMNPQRSTVWY")
}
