% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTools.R
\name{getResidues}
\alias{getResidues}
\title{Get Standard Amino Acid Residue List (Internal)}
\usage{
getResidues()
}
\value{
A character vector of valid residue codes.
}
\description{
Returns the list of amino acid residue codes used internally.
}
\keyword{internal}
