% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Alignment.R
\name{getScore}
\alias{getScore}
\alias{getScore,Alignment-method}
\title{Get Covariance Score}
\usage{
getScore(object, ...)

\S4method{getScore}{Alignment}(object, i, j)
}
\arguments{
\item{object}{An \code{Alignment} object.}

\item{...}{Additional arguments (e.g., residue positions).}

\item{i}{An integer. Index of the first alignment column.}

\item{j}{An integer. Index of the second alignment column.}
}
\value{
A numeric covariance score.

A numeric covariance score between the specified columns.
}
\description{
Generic function to compute a covariance score between residues in an alignment.
}
\section{Functions}{
\itemize{
\item \code{getScore(Alignment)}: Computes the covariance score between two columns in an \code{Alignment} object. Frequencies and residue pairings are calculated to produce a score based on variability and co-occurrence.

}}
