% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getScore3}
\alias{getScore3}
\title{Get Covariance Score (Version 3)}
\usage{
getScore3(object, ...)
}
\arguments{
\item{object}{An \code{Alignment} object.}

\item{...}{Additional arguments.}
}
\value{
A numeric covariance score.
}
\description{
Alternate version of \code{getScore}, that may include perfectly conserved residues.
}
