% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Alignment.R
\name{grepNames}
\alias{grepNames}
\title{Grep Sequence Names from Alignment}
\usage{
grepNames(object, pattern)
}
\arguments{
\item{object}{An \code{Alignment} object.}

\item{pattern}{A regular expression pattern (as a character string).}
}
\value{
A character vector of matching sequence names.
}
\description{
Returns sequence names that match the given regex pattern.
}
