% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{numericORNULL-class}
\alias{numericORNULL-class}
\title{Class Union: numericORNULL (Internal)}
\description{
Internal class union for a slot that can contain either a numeric value or NULL.
}
\keyword{internal}
