% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTools.R
\name{project_geometrical_matrix}
\alias{project_geometrical_matrix}
\title{Project Lower-Triangle Matrix Internally}
\usage{
project_geometrical_matrix(raw_data)
}
\arguments{
\item{raw_data}{A data frame with columns \code{i}, \code{j}, and \code{score}, usually output from \code{getMICovarianceScore}.}
}
\value{
A symmetric data frame with columns \code{i}, \code{j}, and \code{score}.
}
\description{
Used internally to mirror mutual information scores across the diagonal for visualization purposes.
}
\keyword{internal}
