% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Alignment.R
\name{readMask}
\alias{readMask}
\title{Read Alignment Column Mask}
\usage{
readMask(mask_path)
}
\arguments{
\item{mask_path}{Path to the mask file}
}
\value{
An integer vector of column indices
}
\description{
Reads a mask file containing a line like '#ColumnsMap 586, 587, 588, ...'
and returns the list of column indices.
}
