% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Alignment.R
\name{readMICSV}
\alias{readMICSV}
\title{Read MI CSV File}
\usage{
readMICSV(path)
}
\arguments{
\item{path}{Path to the CSV file}
}
\value{
A data frame with columns i, j, score
}
\description{
Reads a mutual information score file with columns i, j, and score.
}
