% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTools.R
\docType{data}
\name{residues}
\alias{residues}
\title{Standard amino acid residue list (internal).}
\format{
An object of class \code{character} of length 20.
}
\usage{
residues
}
\description{
Standard amino acid residue list (internal).
}
\keyword{internal}
