% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTools.R
\docType{data}
\name{residues_length}
\alias{residues_length}
\title{Residue list length (internal).}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
residues_length
}
\description{
Residue list length (internal).
}
\keyword{internal}
